chore: merge babel-config-maker
