+++ babel.config.js
@@ -1,2 +1,26 @@
-const makeBabelConfig = require('./gui-common/babel-config-maker');
+function makeBabelConfig (electronVersion) {
+  return (api) => {
+    api.cache(true);
+    return {
+      plugins: [
+        '@babel/plugin-syntax-jsx',
+        '@babel/plugin-transform-react-jsx',
+        '@babel/plugin-transform-runtime',
+        'transform-class-properties',
+      ],
+      presets: [
+        [
+          '@babel/preset-env',
+          {
+            targets: { electron: electronVersion },
+            useBuiltIns: 'usage',
+            corejs: 2,
+          }
+        ],
+        ['@babel/preset-react'],
+      ],
+    };
+  };
+}
+
 module.exports = makeBabelConfig(require('./node_modules/electron/package.json').version);
